requirements.txt
opencv-python>=4.8.0
ffmpeg-python>=0.2.0
pydub>=0.25.1
PyQt6>=6.6.0
numpy>=1.24.0